+++ core/src/classes/Monk.java
@@ -4,7 +4,7 @@ import utils.Party;
 
 public class Monk extends BaseHero {
 
-    public Monk(Party side, int x, int y, int amount) {
+    public Monk(Party side, float x, float y, int amount) {
         super(12, 7, 30, 5, -4, -4, 1,"Monk", "alive");
         super.side = side;
         super.coords = new CharsCoords(x, y);
