+++ core/src/classes/Peasant.java
@@ -5,7 +5,7 @@ import utils.Party;
 public class Peasant extends BaseHero {
     private final int delivery = 1;
 
-    public Peasant (Party side, int x, int y, int amount) {
+    public Peasant (Party side, float x, float y, int amount) {
         super(1, 1, 1, 3, 1, 1, 1,"Peasant", "alive");
         super.side = side;
         super.coords = new CharsCoords(x, y);
