+++ core/src/classes/Sharpshooter.java
@@ -6,7 +6,7 @@ import utils.Party;
 public class Sharpshooter extends BaseHero {
     private int shots = 32;
 
-    public Sharpshooter(Party side, int x, int y, int amount) {
+    public Sharpshooter(Party side, float x, float y, int amount) {
         super(12, 10, 15, 9, 8, 10, 1,"Sharpshooter", "alive");
         super.side = side;
         super.coords = new CharsCoords(x, y);
