+++ core/src/classes/Xbowman.java
@@ -6,7 +6,7 @@ import utils.Party;
 public class Xbowman extends BaseHero {
     private int shots = 16;
 
-    public Xbowman(Party side, int x, int y, int amount) {
+    public Xbowman(Party side, float x, float y, int amount) {
         super(6, 3, 11, 4, 2, 3, 1,"Xbowman", "alive");
         super.side = side;
         super.coords = new CharsCoords(x, y);
